Move bits of code to their own file to encourage more sensible sharing of utilities
